tests: moved test to correct location
